Lab 8.14.1
Finally completed. It took a while to figure out how to calculate the average of each assignment. I wasn't able to do it only using dictionary methods. I used the value method on the dictionary and then converted that into a list (using the list method). Then I used that list to calculate the average score for each assignment.